+++ src/syntaxDiags.ts
@@ -1,3 +1,3 @@
 import * as vs from 'vscode';
 
-export default vs.languages.createDiagnosticCollection('squirrel');
+export default vs.languages.createDiagnosticCollection('quirrel');